src/requirements.txt
langchain
langchain-openai
langchain-postgres
pypdf
python-dotenv
psycopg
